Remove incorrect ESP32 customization
I2C `ReSTART` operations on ESP32 are queued and not executed until an i2c transaction with a `STOP` is processed. To indicate this queuing, `Wire.endTransmission(false);` or `Wire.requestFrom(id,size,false);` returns `I2C_ERROR_CONTINUE`.  The prior revision of this library added `ReSTART` error checking, BUT, did not use `ReSTART` operations.  So the error detection code was falsely detecting success as an error.